Temporary player stuck fix
There is a bug where the player gets stuck when the chasingenemy attaches to the player. It is now fixed by giving the enemy a brief knockback. However this is not the best fix since the entire game's damage taking is inside the gamescreen class, very hard to customize when player has different weapon or different types of enemies.